d/rules: fix build on ppc64el
Due to a rust bug on ppc64el, `squeekboard-test-layout` fails to build
properly, causing the whole build to fail on this architecture. This
prevents newer versions of the package to migrate to testing.

This bug is due to the linker trying to resolve unused symbols when
building a rust binary and is only seen when using version 1.59.0 of
rustc.

This commit works around this problem by adding the needed linker flags
to ensure all symbols are properly resolved, even those not actually
used by the binaries.

Those changes can be reverted once a newer version of rustc is packaged
for Debian.

Note: this commit also drops the `xgot` feature for mips64el as it is
now automatically enabled for this architecture in Debian's rustc.

Closes: #1016396