us.yaml and us_wide.yaml: Change outline of 123- and ABC-keys to altline
The altline-outline is primarily meant to be used for keys
that change the view, which is what these keys do.

Also changing it on the actions-view would make the layout not align
well anymore, as it is currently not possible to have 2 different
sizes with the altline-styling.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/601>