data/keyboards/de*.yaml: Change äÄ-label to Ää
A big letter after a small letter looks incorrect on a keyboard,
because that is not normal German grammar and a keyboard is for
inputting text.
The key is also on the left side of the layout,
which makes äÄ not fit the visual flow.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/611>