Improve German layout
It is rectangular now,
there are less layout-changes when switching views,
and some commonly used symbols are more easily available.

Before this change, the arrangement of the keys was different between
the normal layout and the wide layout.
The wide layout can now move to the "eschars"-view, like the normal layout.

Part-of: <https://gitlab.gnome.org/World/Phosh/squeekboard/-/merge_requests/627>